+++ hri_safety_sense/include/hri_safety_sense/VehicleInterface.h
