+++ hri_safety_sense/src/MsgHandler.h
@@ -19,7 +19,7 @@
  * Includes
  */
 #include "ros/ros.h"
-#include "VehicleMessages.h"
+#include "hri_safety_sense/VehicleMessages.h"
 
 namespace hri_safety_sense {
